+++ packages/components/psammead-navigation/src/index.stories.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import styled from 'styled-components';
 import { storiesOf } from '@storybook/react';
 import {
   color,
@@ -11,7 +10,6 @@ import {
 } from '@storybook/addon-knobs';
 import * as svgs from '@bbc/psammead-assets/svgs';
 import { C_POSTBOX, C_WHITE } from '@bbc/psammead-styles/colours';
-import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
 import Brand from '@bbc/psammead-brand';
 import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
 import Navigation, { NavigationUl, NavigationLi } from './index';
@@ -25,41 +23,26 @@ import notes from '../README.md';
 const navStoriesData = [
   {
     title: 'igbo',
-    skipLinkText: 'Wụga n’ọdịnaya',
     currentPageText: 'Current page',
     data: igboNavData,
   },
   {
     title: 'pidgin',
-    skipLinkText: 'Waka go wetin de inside',
     currentPageText: 'Current page',
     data: pidginNavData,
   },
   {
     title: 'yoruba',
-    skipLinkText: 'Fò kọjá sí nnkan tí ó wà nínú rẹ̀',
     currentPageText: 'Current page',
     data: yorubaNavData,
   },
   {
     title: 'arabic',
-    skipLinkText: 'إذهب الى المحتوى',
     currentPageText: 'Current page',
     data: arabicNavData,
     dir: 'rtl',
   },
 ];
-
-const offScreenText = (
-  <>
-    <span
-      // eslint-disable-next-line jsx-a11y/aria-role
-      role="text"
-    >
-      <span lang="en-GB">BBC News</span>, Ìgbò - Akụkọ
-    </span>
-  </>
-);
 
 const inputs = () => {
   // capitalization is only for presentation purpose on the knob
@@ -126,10 +109,6 @@ const getBrand = () => {
   );
 };
 
-const StyledMain = styled.main`
-  padding: 0px 1rem;
-`;
-
 const navigationStory = (
   skipLinkText,
   currentPageText,
@@ -166,11 +145,6 @@ const navigationStory = (
         })}
       </NavigationUl>
     </Navigation>
-    <StyledMain>
-      <VisuallyHiddenText id="content" as="h1" tabIndex="-1">
-        {offScreenText}
-      </VisuallyHiddenText>
-    </StyledMain>
   </>
 );